requirements.txt
Django==1.10.3
django-debug-toolbar==1.6
django-tastypie==0.13.3
mysqlclient==1.3.9
python-dateutil==2.6.0
python-mimeparse==1.6.0
six==1.10.0
sqlparse==0.2.2
